add kifuable-parser
