+++ lib/jkf/parser/kif.rb
@@ -2,6 +2,8 @@
 
 module Jkf::Parser
   class Kif < Base
+    include Kifuable
+
     def parse_root
       @input += "\n" unless @input.end_with?("\n")
 
@@ -178,7 +180,7 @@ module Jkf::Parser
           @current_pos = s4
         end
         @reported_pos = s0
-        s0 = transform_initial_board(s3)
+        s0 = transform_initialboard(s3)
       else
         @current_pos = s0
         s0 = :failed
@@ -1011,10 +1013,6 @@ module Jkf::Parser
       s0
     end
 
-    def parse_nonl
-      match_regexp(/^[^\r\n]/)
-    end
-
     protected
 
     def transform_root(headers, ini, headers2, moves, forks)
@@ -1027,56 +1025,12 @@ module Jkf::Parser
         preset = preset2str(ret["header"]["手合割"])
         ret["initial"] = { "preset" => preset } if preset && preset != "OTHER"
       end
-      transform_root_header_data(ret)
+      transform_root_header_data(ret) if ret["initial"] && ret["initial"]["data"]
       transform_root_forks(forks, moves)
       if ret["initial"] && ret["initial"]["data"] && ret["initial"]["data"]["color"] == 1
         reverse_color(ret["moves"])
       end
       ret
-    end
-
-    def transform_root_header_data(ret)
-      if ret["initial"] && ret["initial"]["data"]
-        if ret["header"]["手番"]
-          ret["initial"]["data"]["color"] = "下先".include?(ret["header"]["手番"]) ? 0 : 1
-          ret["header"].delete("手番")
-        else
-          ret["initial"]["data"]["color"] = 0
-        end
-        ret["initial"]["data"]["hands"] = [
-          make_hand(ret["header"]["先手の持駒"] || ret["header"]["下手の持駒"]),
-          make_hand(ret["header"]["後手の持駒"] || ret["header"]["上手の持駒"])
-        ]
-        %w(先手の持駒 下手の持駒 後手の持駒 上手の持駒).each do |key|
-          ret["header"].delete(key)
-        end
-      end
-    end
-
-    def transform_root_forks(forks, moves)
-      fork_stack = [{ "te" => 0, "moves" => moves }]
-      forks.each do |f|
-        now_fork = f
-        _fork = fork_stack.pop
-        _fork = fork_stack.pop while _fork["te"] > now_fork["te"]
-        move = _fork["moves"][now_fork["te"] - _fork["te"]]
-        move["forks"] ||= []
-        move["forks"] << now_fork["moves"]
-        fork_stack << _fork
-        fork_stack << now_fork
-      end
-    end
-
-    def transform_initial_board(lines)
-      ret = []
-      9.times do |i|
-        line = []
-        9.times do |j|
-          line << lines[j][8 - i]
-        end
-        ret << line
-      end
-      { "preset" => "OTHER", "data" => { "board" => ret } }
     end
 
     def transform_move(line, c)
@@ -1103,54 +1057,6 @@ module Jkf::Parser
       ret
     end
 
-    def zen2n(s)
-      "０１２３４５６７８９".index(s)
-    end
-
-    def kan2n(s)
-      "〇一二三四五六七八九".index(s)
-    end
-
-    def kan2n2(s)
-      case s.length
-      when 1
-        "〇一二三四五六七八九十".index(s)
-      when 2
-        "〇一二三四五六七八九十".index(s[1]) + 10
-      else
-        raise "21以上の数値に対応していません"
-      end
-    end
-
-    def kind2csa(kind)
-      if kind[0] == "成"
-        {
-          "香" => "NY",
-          "桂" => "NK",
-          "銀" => "NG"
-        }[kind[1]]
-      else
-        {
-          "歩" => "FU",
-          "香" => "KY",
-          "桂" => "KE",
-          "銀" => "GI",
-          "金" => "KI",
-          "角" => "KA",
-          "飛" => "HI",
-          "玉" => "OU",
-          "王" => "OU",
-          "と" => "TO",
-          "杏" => "NY",
-          "圭" => "NK",
-          "全" => "NG",
-          "馬" => "UM",
-          "竜" => "RY",
-          "龍" => "RY"
-        }[kind]
-      end
-    end
-
     def special2csa(str)
       {
         "中断" => "CHUDAN",
@@ -1163,26 +1069,6 @@ module Jkf::Parser
         "反則勝ち" => "ILLEGAL_ACTION", # 直前の手が反則(先頭に+か-で反則した側の情報を含める必要が有る)
         "反則負け" => "ILLEGAL_MOVE" # ここで手番側が反則，反則の内容はコメントで表現
       }[str] || (raise ParseError)
-    end
-
-    def preset2str(preset)
-      {
-        "平手" => "HIRATE",
-        "香落ち" => "KY",
-        "右香落ち" => "KY_R",
-        "角落ち" => "KA",
-        "飛車落ち" => "HI",
-        "飛香落ち" => "HIKY",
-        "二枚落ち" => "2",
-        "三枚落ち" => "3",
-        "四枚落ち" => "4",
-        "五枚落ち" => "5",
-        "左五枚落ち" => "5_L",
-        "六枚落ち" => "6",
-        "八枚落ち" => "8",
-        "十枚落ち" => "10",
-        "その他" => "OTHER"
-      }[preset.gsub(/\s/, "")]
     end
 
     def teban2color(teban)
